+++ tests/api-resources/api/subscriptions/subscriptions.test.ts
@@ -67,6 +67,7 @@ describe('resource subscriptions', () => {
       delivery_address_id: 'delivery_address_id',
       delivery_slot_id: 'delivery_slot_id',
       note: 'note',
+      notify_customer: true,
     });
   });
 
@@ -143,7 +144,11 @@ describe('resource subscriptions', () => {
     await expect(
       client.api.subscriptions.cancel(
         'id',
-        { cancel_immediately: true, reason: 'reason' },
+        {
+          cancel_immediately: true,
+          reason: 'reason',
+          silent: true,
+        },
         { path: '/_stainless_unknown_path' },
       ),
     ).rejects.toThrow(Gro.NotFoundError);
@@ -315,7 +320,11 @@ describe('resource subscriptions', () => {
 
   // Mock server tests are disabled
   test.skip('skip: required and optional params', async () => {
-    const response = await client.api.subscriptions.skip('id', { skip_weeks: 0, reason: 'reason' });
+    const response = await client.api.subscriptions.skip('id', {
+      skip_weeks: 0,
+      reason: 'reason',
+      skip_reason_id: 'skip_reason_id',
+    });
   });
 
   // Mock server tests are disabled